# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Estructura de las funciones de la API de Iglu's Shop",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Crear un nuevo producto",
                "description": "Crea un nuevo producto con los datos proporcionados. \n\nRequiere autenticación y permisos de administrador.\n\nAl crear un nuevo producto, se deben proporcionar los siguientes campos obligatorios:\n\n- `name`: El nombre del producto, que debe ser una cadena de texto con un mínimo de 3 caracteres.\n- `description`: Descripción detallada del producto, la cual debe ser una cadena de texto con un máximo de 255 caracteres.\n- `price`: El precio del producto, que debe ser un número decimal con un máximo de dos decimales.\n- `category_id`: ID de la categoría a la que pertenece el producto. Debe ser un número entero y debe existir en la base de datos.\n- `provider_id`: ID del proveedor asociado al producto. Debe ser un número entero y debe existir en la base de datos.\n\nLos siguientes campos son opcionales, pero si se incluyen, deben cumplir con las siguientes validaciones:\n\n- `stock`: Número entero que indica la cantidad disponible del producto en stock.\n- `image`: URL válida que apunta a la imagen del producto.\n- `has_discount`: Campo booleano que indica si el producto tiene descuento. Este campo debe ser siempre falso para los productos estándar. Si se quiere generar un producto \n   con descuento se debe hacer desde el endpoint correspondiente de outlet.\n- `discount`: El descuento aplicado al producto. Si `has_discount` es falso, este valor debe ser 0.\n\nLa validación incluye restricciones de tipo de datos y valores específicos. Si cualquier campo no cumple con las reglas, se devolverá un error.",
                "operationId": "e8185e307706e2cd84bbcf5171e2f09d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto creado correctamente."
                                        },
                                        "id": {
                                            "type": "integer",
                                            "example": 21
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obtener un producto por ID",
                "description": "Devuelve los datos de un producto específico.\n\nEs un endpoint público \n\nBusca el producto por su ID y comprueba que no pertenezca al outlet.\nSi no se encuentra, devuelve un error",
                "operationId": "7ab21ddb713501f3d4c2914892c02a21",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Producto no encontrado o perteneciente al outlet",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Actualizar un producto",
                "description": "Actualiza los datos de un producto existente ya sea del outlet o no.\n\nRequiere autenticación y permisos de administrador.  \n\nPermite actualizar uno o varios campos del producto. \nDevuelve el producto actualizado si todo es correcto. \n\nPermite actualizar el campo has_discount a un valor truthy para pasar el producto al \nendpoint de outlet",
                "operationId": "6a7e11900844cd85cff66e662b82bd4c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto actualizado correctamente"
                                        },
                                        "product": {
                                            "$ref": "#/components/schemas/ProductUpdate"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Elimina un producto",
                "description": "Elimina un producto que no pertenezca al outlet por su ID. \n\nRequiere autenticación y permisos de administrador.\n\nSi el producto no existe, devuelve un error 404. \nSi se elimina correctamente, devuelve un código 204.",
                "operationId": "e80a6ca46716acee6b47ea1ac91663c0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del producto a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Producto eliminado correctamente"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "Producto no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "NotFoundError": {
                "title": "Recurso no encontrado",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Usuario no encontrado"
                    }
                },
                "type": "object"
            },
...
            "ProductCreate": {
                "required": [
                    "name",
                    "description",
                    "price",
                    "category_id",
                    "provider_id"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 3,
                        "example": "Camiseta Deportiva"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Camiseta técnica de alto rendimiento"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 29.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 100
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/product.jpg"
                    },
                    "has_discount": {
                        "description": "Debe ser false o equivalente (declined). El producto creado NO puede tener descuento inicialmente.",
                        "type": "boolean",
                        "enum": [
                            false
                        ],
                        "example": false
                    },
                    "discount": {
                        "description": "Debe ser 0 si no hay descuento.",
                        "type": "number",
                        "example": 0
                    },
                    "category_id": {
                        "description": "ID de la categoría relacionada (clave foránea a Category)",
                        "type": "integer",
                        "example": 1
                    },
                    "provider_id": {
                        "description": "ID del proveedor relacionada (clave foránea a Provider)",
                        "type": "integer",
                        "example": 5
                    }
                },
                "type": "object"
            },
            "Product": {
                "title": "Product",
                "required": [
                    "id",
                    "name",
                    "description",
                    "price",
                    "stock",
                    "image",
                    "has_discount",
                    "discount",
                    "provider_id",
                    "category_id",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Camiseta Deportiva"
                    },
                    "description": {
                        "type": "string",
                        "example": "Camiseta técnica de alto rendimiento"
                    },
                    "price": {
                        "type": "number",
                        "example": 29.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 2
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://www.ejemplo.com/imagenes/camiseta.jpg"
                    },
                    "has_discount": {
                        "description": "Este campo debe ser siempre false. si se cambia a true el producto pasa al outlet",
                        "type": "boolean",
                        "default": 0,
                        "example": false
                    },
                    "discount": {
                        "description": "Si has_discount es falsy entonces discount debe ser 0",
                        "type": "number",
                        "example": 0
                    },
                    "provider_id": {
                        "description": "ID del proveedor relacionada (clave foránea a Provider)",
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "description": "ID de la categoría relacionada (clave foránea a Category)",
                        "type": "integer",
                        "example": 1
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-05 21:45:59"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-04-05 21:45:59"
                    }
                },
                "type": "object"
            },
            "ProductUpdate": {
                "title": "Actualización de Producto",
                "description": "Schema para actualizar un producto",
                "properties": {
                    "name": {
                        "type": "string",
                        "minLength": 3,
                        "example": "Camiseta Deportiva"
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Camiseta cómoda para entrenamiento"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 49.99
                    },
                    "stock": {
                        "type": "integer",
                        "example": 150
                    },
                    "image": {
                        "type": "string",
                        "format": "uri",
                        "example": "https://example.com/images/producto.png"
                    },
                    "has_discount": {
                        "type": "boolean",
                        "example": false
                    },
                    "discount": {
                        "type": "number",
                        "example": 0
                    },
                    "category_id": {
                        "description": "ID de la categoría relacionada (clave foránea a Category)",
                        "type": "integer",
                        "example": 1
                    },
                    "provider_id": {
                        "description": "ID del proveedor relacionada (clave foránea a Provider)",
                        "type": "integer",
                        "example": 2
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Error de validación",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "titulo": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "El nombre es obligatorio."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}